% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA earthquake data}
\usage{
eq_clean_data(path_of_data = "inst/extdata/earthquakes.tsv")
}
\arguments{
\item{path_of_data}{character string containing path of .tsv file for NOAA earthquake data.}
}
\value{
A clean data frame with date, latitude, and longitude columns.
}
\description{
This function takes raw NOAA earthquake data and returns a clean data frame.
The clean data frame includes a date column created by uniting the year, month, and day,
converting it to the Date class. It also converts the LATITUDE and LONGITUDE columns to numeric class.
}
\examples{
\dontrun{
data <- read.csv(system.file("extdata", "your_data.csv", package = "visualizeNOAA"))
eq_clean_data(data)
}

}
