% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create HTML Labels for Leaflet Map Annotations}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A dataset containing earthquake information.}
}
\value{
A character vector with HTML labels for each earthquake.
}
\description{
This function takes a dataset as an argument and creates an HTML label that can be
used as the annotation text in a leaflet map. The function puts together a character
string for each earthquake that will show the cleaned location, the magnitude, and
the total number of deaths, with boldface labels for each ("Location", "Total deaths",
and "Magnitude").
}
\examples{
\dontrun{
data <- read.csv("earthquake_data.csv")
eq_create_label(data)
}

}
