% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean earthquake location names}
\usage{
eq_location_clean(data_df)
}
\arguments{
\item{data_df}{A dataframe containing raw earthquake data with a LOCATION_NAME column.}
}
\value{
A dataframe with the LOCATION_NAME column cleaned.
}
\description{
This function cleans the LOCATION_NAME column by stripping out the country name
(including the colon) and converts names to title case (as opposed to all caps).
}
