% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Geom for plotting a time line of earthquakes}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mappings created by \code{\link[=aes]{aes()}} or \code{\link[=aes_]{aes_()}}.}

\item{data}{The data to be displayed in this layer.}

\item{stat}{The statistical transformation to use on the data for this layer.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{Logical. Should this layer be included in the legends?}

\item{inherit.aes}{If \code{TRUE}, inherits aesthetics from the plot default,
which is set with \code{\link[=theme_set]{theme_set()}}. If \code{FALSE}, overrides all aesthetics.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer}}.
These are often aesthetics, used to set an aesthetic to a fixed value, like \code{color = "red"} or \code{size = 3}.
They may also be parameters to the paired geom/stat.}
}
\description{
Build a geom for ggplot2 called geom_timeline() for plotting a time line of earthquakes
ranging from \code{xmin} to \code{xmax} dates with a point for each earthquake.
}
