% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Geom for plotting a time line of earthquakes}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomTimeline
}
\description{
Build a geom for ggplot2 called GeomTimeline for plotting a time line of earthquakes
ranging from \code{xmin} to \code{xmax} dates with a point for each earthquake.
}
\keyword{datasets}
