% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_code.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Geom for adding annotations to earthquake data}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomTimelineLabel
}
\description{
Build a geom called GeomTimelineLabel for adding annotations to the earthquake data.
This geom adds a vertical line to each data point with a text annotation (e.g., the location
of the earthquake) attached to each line. There should be an option to subset to \code{n_max}
number of earthquakes, where we take the \code{n_max} largest (by magnitude) earthquakes.
}
\keyword{datasets}
